// irmin-console/src/types/Wordpress.ts
export type Menu = MenuItem[];

export interface MenuItem {
  ID: number;
  title: string;
  url: string;
  menu_order: number;
  menu_item_parent: string;
}